fixed and added tests
